coVRNavigationManager: extract coVRNavigationProvider class
