+++ src/OpenCOVER/cover/CMakeLists.txt
@@ -48,6 +48,7 @@ SET(COVER_HEADERS
   coVRMessageSender.h
   coVRMSController.h
   coVRNavigationManager.h
+  coVRNavigationProvider.h
   coVRPartner.h
   coVRPlugin.h
   coVRPluginList.h
@@ -119,6 +120,7 @@ SET(COVER_SOURCES
   coVRMessageSender.cpp
   coVRMSController.cpp
   coVRNavigationManager.cpp
+  coVRNavigationProvider.cpp
   coVRPartner.cpp
   coVRPlugin.cpp
   coVRPluginList.cpp
