+++ src/OpenCOVER/cover/coVRNavigationManager.cpp
@@ -1,4 +1,4 @@
-﻿/* This file is part of COVISE.
+/* This file is part of COVISE.
 
    You can use it under the terms of the GNU Lesser General Public License
    version 2.1 or later, see lgpl-2.1.txt.
@@ -114,21 +114,6 @@ static float mouseScreenWidth()
 static float mouseScreenHeight()
 {
     return Input::instance()->mouse()->screenHeight();
-}
-
-coVRNavigationProvider::coVRNavigationProvider(const std::string n, coVRPlugin* p)
-{
-    name = n;
-    plugin = p;
-    ID = -1;
-}
-coVRNavigationProvider::~coVRNavigationProvider()
-{
-}
-
-void coVRNavigationProvider::setEnabled(bool state)
-{
-    enabled = state;
 }
 
 coVRNavigationManager *coVRNavigationManager::instance()
